Add files via upload
When attempting to update the tunnel configuration with the tunnels connected, the update would fail with "The VPN connection tunnelname cannot be removed from the global user connections. : Cannot delete a connection while it is connected".
The script will now disconnect all active network adapters in the system to terminate the VPN connection, update the configuration and enable any disabled adapters.